The error is no more, and fixed some other errors
While fixing the error I encountered some others so I fixed them.
Now unless I get to catch another, the code is perfect for simple
operations. Less gooooo!!!

+++ lpp/lexer.cpp
@@ -70,7 +70,7 @@
     std::vector<Token> Lexer::processResult(std::vector<Token> vec) // TODO finish the method
     {
         position = 0;
-        while (position < vec.size())
+        while (position < (int)vec.size())
         {
             if (!vec[position].value.compare("<") && !vec[position + 1].value.compare("="))
             {
